requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
transformers==4.38.2
spacy==3.7.4
aiohttp==3.9.3
python-dotenv==1.0.1
jinja2==3.1.3 